requirements.txt
asgiref==3.3.1
atomicwrites==1.4.0
attrs==19.3.0
certifi==2019.9.11
cffi==1.14.4
chardet==3.0.4
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.2.1
defusedxml==0.7.0rc1
Django==3.1.3
django-cors-headers==3.5.0
django-templated-mail==1.1.1        
djangorestframework==3.11.0
djangorestframework-simplejwt==4.3.0
djoser==2.1.0
gunicorn==20.0.4
idna==2.8
importlib-metadata==1.5.0
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
more-itertools==8.2.0
oauthlib==3.1.0
packaging==20.1
pluggy==0.13.1
psycopg2-binary==2.8.5
py==1.8.1
pycparser==2.20
PyJWT==1.7.1
pyparsing==2.4.6
pytest==5.3.5
pytest-django==3.8.0
python3-openid==3.2.0
pytz==2021.1
requests==2.22.0
requests-oauthlib==1.3.0
six==1.14.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sorl-thumbnail==12.6.3
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.25.6
wcwidth==0.1.8
zipp==2.2.0